+++ tests/RouterTest.php
@@ -149,3 +149,87 @@ it("should pass parameters correctly if the URL matches", function (): void {
 
     expect(Router::matchRequestedUrl())->toBe("showing post #42");
 });
+
+it("should get an URL by the route name", function (): void {
+    Router::addGetRoute("/about-us", function (): void {
+    }, "about-us.index");
+
+    expect(Router::getRouteUrl("about-us.index"))->toBe("/about-us");
+});
+
+it("should get an URL by its route name even if it has parameters, and parameters are filled by index", function (): void {
+    Router::addGetRoute("/user/{user}/post/{post}", function (): void {
+    }, "user.post.show");
+
+    expect(Router::getRouteUrl("user.post.show", [1, 42]))->toBe("/user/1/post/42");
+});
+
+it("should get an URL by its route name even if it has parameters, and parameters are filled by key name", function (): void {
+    Router::addGetRoute("/user/{user}/post/{post}", function (): void {
+    }, "user.post.show");
+
+    expect(Router::getRouteUrl("user.post.show", ["user" => 1, "post" => 42]))->toBe("/user/1/post/42");
+});
+
+it("should get an URL by its route name event if it has parameters, and parameters are filled by key name and index", function (): void {
+    Router::addGetRoute("/user/{user}/post/{post}", function (): void {
+    }, "user.post.show");
+
+    expect(Router::getRouteUrl("user.post.show", ["user" => 1, 42]))->toBe("/user/1/post/42");
+});
+
+it("should get the URL of a parameterized POST route", function (): void {
+    Router::addPostRoute("/user/{user}/post", function (): void {
+    }, "user.post.store");
+
+    expect(Router::getRouteUrl("user.post.store", ["user" => 42]))->toBe("/user/42/post");
+});
+
+it("should throw an exception if the route name does not exist", function (): void {
+    $this->expectException(OutOfRangeException::class);
+
+    Router::getRouteUrl("home.index");
+});
+
+it("should throw an exception if the route name is empty", function (): void {
+    $this->expectException(InvalidArgumentException::class);
+
+    Router::addGetRoute("/", function (): void {
+    }, "");
+});
+
+it("should throw an exception if one of the parameter is missing", function (): void {
+    $this->expectException(InvalidArgumentException::class);
+
+    Router::addGetRoute("/user/{user}/post/{post}", function (): void {
+    }, "user.post.show");
+
+    Router::getRouteUrl("user.post.show", ["user" => 42]);
+});
+
+it("should throw an exception message if one of the parameter is missing", function (): void {
+    $this->expectExceptionMessage("parameter post missing for route /user/{user}/post/{post}");
+
+    Router::addGetRoute("/user/{user}/post/{post}", function (): void {
+    }, "user.post.show");
+
+    Router::getRouteUrl("user.post.show", ["user" => 42]);
+});
+
+it("should throw an exception if the parameter does not match the regexp", function (): void {
+    $this->expectException(InvalidArgumentException::class);
+
+    Router::addGetRoute("/user/{user}", function (): void {
+    }, "user.show");
+
+    Router::getRouteUrl("user.show", ["/42"]);
+});
+
+it("should throw an exception message if the parameter does not match the regexp", function (): void {
+    $this->expectExceptionMessage("parameter user does not matches regular expression [^/]+ for route /user/{user}");
+
+    Router::addGetRoute("/user/{user}", function (): void {
+    }, "user.show");
+
+    echo Router::getRouteUrl("user.show", ["/42"]);
+});